feat(sdk-metrics-base): document and export basic APIs
