+++ experimental/packages/opentelemetry-sdk-metrics-base/src/InstrumentDescriptor.ts
@@ -15,9 +15,20 @@
  */
 
 import { MetricOptions, ValueType } from '@opentelemetry/api-metrics-wip';
-import { InstrumentType } from './Instruments';
 import { View } from './view/View';
 
+/**
+ * Supported types of metric instruments.
+ * @see https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument
+ */
+ export enum InstrumentType {
+  COUNTER = 'COUNTER',
+  HISTOGRAM = 'HISTOGRAM',
+  UP_DOWN_COUNTER = 'UP_DOWN_COUNTER',
+  OBSERVABLE_COUNTER = 'OBSERVABLE_COUNTER',
+  OBSERVABLE_GAUGE = 'OBSERVABLE_GAUGE',
+  OBSERVABLE_UP_DOWN_COUNTER = 'OBSERVABLE_UP_DOWN_COUNTER',
+}
 
 export interface InstrumentDescriptor {
   readonly name: string;
