+++ experimental/packages/opentelemetry-sdk-metrics-base/src/Instruments.ts
@@ -19,17 +19,6 @@ import * as metrics from '@opentelemetry/api-metrics-wip';
 import { InstrumentDescriptor } from './InstrumentDescriptor';
 import { WritableMetricStorage } from './state/WritableMetricStorage';
 
-// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument
-
-export enum InstrumentType {
-  COUNTER = 'COUNTER',
-  HISTOGRAM = 'HISTOGRAM',
-  UP_DOWN_COUNTER = 'UP_DOWN_COUNTER',
-  OBSERVABLE_COUNTER = 'OBSERVABLE_COUNTER',
-  OBSERVABLE_GAUGE = 'OBSERVABLE_GAUGE',
-  OBSERVABLE_UP_DOWN_COUNTER = 'OBSERVABLE_UP_DOWN_COUNTER',
-}
-
 export class SyncInstrument {
   constructor(private _writableMetricStorage: WritableMetricStorage, private _descriptor: InstrumentDescriptor) { }
 
@@ -42,13 +31,27 @@ export class SyncInstrument {
   }
 }
 
-export class UpDownCounter extends SyncInstrument implements metrics.Counter {
+/**
+ * The class implements {@link metrics.UpDownCounter} interface.
+ * @see https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#updowncounter
+ */
+export class UpDownCounter extends SyncInstrument implements metrics.UpDownCounter {
+  /**
+   * Increment value of counter by the input. Inputs may be negative.
+   */
   add(value: number, attributes?: metrics.Attributes, ctx?: api.Context): void {
     this.aggregate(value, attributes, ctx);
   }
 }
 
+/**
+ * The class implements {@link metrics.Counter} interface.
+ * @see https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#counter
+ */
 export class Counter extends SyncInstrument implements metrics.Counter {
+  /**
+   * Increment value of counter by the input. Inputs may not be negative.
+   */
   add(value: number, attributes?: metrics.Attributes, ctx?: api.Context): void {
     if (value < 0) {
       api.diag.warn(`negative value provided to counter ${this.getName()}: ${value}`);
@@ -59,7 +62,14 @@ export class Counter extends SyncInstrument implements metrics.Counter {
   }
 }
 
+/**
+ * The class implements {@link metrics.Histogram} interface.
+ * @see https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#histogram
+ */
 export class Histogram extends SyncInstrument implements metrics.Histogram {
+  /**
+   * Records a measurement. Value of the measurement must not be negative.
+   */
   record(value: number, attributes?: metrics.Attributes, ctx?: api.Context): void {
     this.aggregate(value, attributes, ctx);
   }
