+++ experimental/packages/opentelemetry-sdk-metrics-base/test/util.ts
@@ -18,8 +18,8 @@ import { Attributes, ValueType } from '@opentelemetry/api-metrics';
 import { InstrumentationLibrary } from '@opentelemetry/core';
 import { Resource } from '@opentelemetry/resources';
 import * as assert from 'assert';
-import { InstrumentDescriptor } from '../src/InstrumentDescriptor';
-import { Histogram, InstrumentType } from '../src/Instruments';
+import { InstrumentDescriptor, InstrumentType } from '../src/InstrumentDescriptor';
+import { Histogram } from '../src/Instruments';
 import { MetricData, PointData, PointDataType } from '../src/export/MetricData';
 import { Measurement } from '../src/Measurement';
 import { isNotNullish } from '../src/utils';
